requirements.txt
argh=0.26.2
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cx-Oracle==7.2.2
dataclasses==0.6
Django==2.2.4
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-jsonview==1.2.0
django-livereload-server==0.3.2
django-livesync==0.4
django-oauth-toolkit==1.2.0
django-pyodbc-azure==2.1.0.0
django-watchman==0.17.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djongo==1.2.33
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
pathtools==0.1.2
Pillow==6.1.0
pip==19.2.3
psycopg2==2.8.3
pymongo==3.9.0
pyodbc==4.0.27
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
setuptools==41.1.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.2.4
tornado==4.5.1
urllib3==1.25.3
watchdog==0.9.0
websocket-client==0.56.0
wheel==0.33.6
whitenoise==4.1.3
gunicorn==19.9.0
python-decouple==3.1
dj-database-url==0.5.0
